added follow parameter and removed client sockets
By using `json-seq`, long-lived connections can be opened to the snapd
requests API so that prompt clients will be automatically notified when
a new request comes in.  This can be done by passing the `follow=true`
option in the query arguments.  In this case, it is no longer necessary
for snapd to send information to a client socket in order to tell the
client to query the snapd API for request information.

Thus, registering a client socket with snapd is also no longer required.
Instead, snapd must be able to identify whether a client should be able
to connect to the snapd request endpoint, and only allow appropriate
clients to be able to connect.

If a client, such as a settings application which is designed to modify
stored decisions, wants to monitor changes to stored decisions without
repeatedly querying the snapd decision API, it can use the same
`follow=true` option in the query to open a long-lived connection in the
same way.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>