Cleaned up `name` -> `snap` + `app` schemas
Removed remaining references to the `name` schema and replaced them with
the new `snap` and `app` schemas.

There still exists a `name` parameter in the
`/v2/prompting/decisions/{name}`, endpoint.  The schema for this name
parameter has been written out explicitly (rather than using the removed
`name` schema).

It should be decided if/how this endpoint should be modified.
Currently, it aligns with the use of `.../{name}` in other endpoints of
the snapd REST API, but not with the way prompting generally works.  In
particular, decisions are stored per-application, rather than per-snap,
while applications are recorded using the snap name appended to the app
name to prevent app name conflicts.  While it may be useful to get all
decisions for a given snap (such as to delete them when uninstalling the
snap with the `--purge` option), it is probably more useful to expose
the stored decisions using the `snap` and `app` names, as that is how
requests are formatted.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>