Replaced decisions `name` endpoint with `snap` and `snap/app`
This change removes the last remaining endpoint using the `name`
parameter.  The use of `name` is similar to that in the snapd REST API,
but it had been assumed to refer to the snap "label" in the context of
prompting, since decisions are stored per-label, e.g. `"snapd.lxd.lxd"`.

Prompt requests are sent using the snap name and app name as separate
parameters so that the prompt client need not parse labels into these
separate components.  Thus, it makes sense for the endpoint parameters
to also follow this pattern, and allow querying all stored decisions for
a given snap or for a given app within a given snap.

Additionally, this change introduces new endpoints to delete all stored
decisions for a given snap or snap+app, which could be useful if a
settings client wants to reset permissions for a given application or if
a snap is uninstalled with the `--purge` option.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>