modified `decisions/{id}` to `decision-id/{id}`
Previously, both the `/v2/prompting/decisions/{id}` and
`/v2/prompting/decisions/{snap}` endpoints shared the same path
preceding the parameter, so it was ambiguous whether the path parameter
was a snap name or decision ID, and thus which endpoint was intended.

This change solves the problem by moving the decision-with-ID endpoint
to a different `decision-id/{id}` path, thus disambiguating the
previously conflicting paths.

Unfortunately, this means decisions-by-ID are on a fundamentally
different path than the other decisions.

The alternative could be to use:

- `/v2/prompting/decisions/by-snap/{snap}`
- `/v2/prompting/decisions/by-snap/{snap}/{app}`
- `/v2/prompting/decisions/by-id/{id}`

but this seems clumsier than simply moving the decisions by ID to their
own endpoint.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>