added `changed-decisions` to response content after `POST`ing response
Responding to a request via sending a `POST` message to
`/v2/prompting/requests/{id}` may or may not result in the creation of a
new rule.  If it does, that rule may modify or delete previous rules
which conflict.  Additionally, the rule which is created may differ from
the original request (in particular, the `permissions`), since some
permissions may be implied by existing rules while others are not.

Thus, by including both the new rule (if it exists) and any modified or
deleted rules, the client is given all relevant information which
changed as a result of the response.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>