Clarified language regarding prompt requests and access decisions.
Previously, the phrase "resource access request" was used to refer to
requests from the kernel to snapd as well as communication between snapd
and the prompt UI client(s), and "resource access request decision" was
used to refer to stored access decisions.

This change makes each of these more clearly distinct, including using a
different schema for request IDs and decision IDs.  In particular, a
"prompt request" now refers to the information communicated between
snapd and the prompt UI client, and an "access decision" refers to the
stored decisions regarding access, which are used internally and may be
communicated to a control center application which can view and modify
access decisions.  This was done for several reasons.

1. Access decisions can be created either via a reply to a prompt
   request or by directly POSTing to the `/v2/prompting/decisions`
   endpoint, so many access decisions do not correspond to any prompt
   request.
2. Stored access decisions are often modified as the result of pruning
   or consolidation when another decision is modified in some way, so
   the contents of the access decision may easily diverge from the
   original contents of the prompt request from which it was created.
3. If IDs are common between prompt requests and access decisions, then
   the system which assigns prompt request IDs must be aware of the
   internal state of the stored access decisions database so as to not
   create a new ID which conflicts with an existing ID.

Additionally, this change replaces the use of the word "resource" with
the phrase "object of request", in reference to the cybersecurity
access control concepts of "subject" and "object".

The hope is that these name changes reinforce the semantic differences
between prompt requests and access decisions, make it clearer that the
former concern prompt UI clients while the latter concern control center
applications, and reduce confusion with other concepts such as HTTP
requests and responses, resources, and apparmor "rules" (and "access" to
a lesser extent).

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>