match snapd API, include path-pattern in reply, and more changes
The existing snapd API uses the `POST` method with an `action` key
instead of using different `DELETE` or `PATCH` methods, for example.
This commit combines the previous `DELETE` operation into the `POST`
operation for those same endpoints.

Additionally, various improvements were made to allow adding multiple
rules at once or deleting the rules for multiple snaps/apps at once.

When multiple prompt requests or access rules are included in a request
or response body, those entries are now in a map with the key being the
request or rule ID, respectively.  This uses a `patternProperties` field
of the openAPI spec which allows for defining arbitrary field names so
long as they match a given regular expression.  If in the future a
format for IDs is decided upon, these patterns could be updated to be
more strict.

The prompting API and backend are moving towards prompt UI clients
returning a path glob as a pattern against which requested paths can be
matched, so this was integrated into the `reply` schema, and the
`path-scope` schema was removed.

Also, IDs were removed from the response bodies returned by operations
which have the ID as a path parameter to the endpoint.  This also
matches the behavior of requests and rules as values in maps where the
key is their ID.

Lastly (I think these are all the major changes), the `allow` schema was
replaced by an `action` schema for use inside of access rules, with the
possible values of `permit` or `deny`.  This has the potential to cause
some confusion with the `action` key in the request body, but the
`action` schema only appears with these semantics inside of `reply` and
`rule-contents` (and as a result, `rule`) schemas.

The README is now out of date, but the snapd Apparmor Prompting spec
SD121 is now up to date with much more detailed information, so I am
inclined to remove the "human-readable" information from the README and
instead encourage Canonical employees to look at SD121 and others to
explore the API spec from within https://editor-next.swagger.io/.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>